Issue 46943: Mothership submission throttling is too aggressive
